Extracted a new interface that can be benchmarked. Simplified the implementation.
