+++ TryAtSoftware.CleanTests.Core/XUnit/CleanTestCaseData.cs
@@ -2,6 +2,7 @@
 
 using System;
 using System.Collections.Generic;
+using TryAtSoftware.CleanTests.Core.Dependencies;
 using TryAtSoftware.Extensions.Collections;
 
 public class CleanTestCaseData
