+++ TryAtSoftware.CleanTests.Core/XUnit/Serialization/SerializableIndividualDependencyNode.cs
@@ -2,6 +2,7 @@
 
 using System;
 using System.Linq;
+using TryAtSoftware.CleanTests.Core.Dependencies;
 using TryAtSoftware.CleanTests.Core.Extensions;
 using TryAtSoftware.Extensions.Collections;
 using Xunit.Abstractions;
